fix Job processor and refactor
