+++ rabix-bindings/src/main/java/org/rabix/bindings/protocol/draft2/Draft2Bindings.java
@@ -20,7 +20,6 @@ import org.rabix.bindings.model.FileValue;
 import org.rabix.bindings.model.Job;
 import org.rabix.bindings.model.dag.DAGNode;
 import org.rabix.bindings.model.requirement.Requirement;
-import org.rabix.bindings.protocol.draft2.helper.Draft2AppProcessor;
 import org.rabix.bindings.protocol.draft2.resolver.Draft2DocumentResolver;
 
 public class Draft2Bindings implements Bindings {
