+++ rabix-bindings/src/main/java/org/rabix/bindings/protocol/draft2/Draft2ValueProcessor.java
@@ -7,7 +7,6 @@ import org.rabix.bindings.ProtocolValueProcessor;
 import org.rabix.bindings.model.FileValue;
 import org.rabix.bindings.model.Job;
 import org.rabix.bindings.protocol.draft2.bean.Draft2Job;
-import org.rabix.bindings.protocol.draft2.helper.Draft2AppProcessor;
 import org.rabix.bindings.protocol.draft2.processor.Draft2PortProcessorException;
 import org.rabix.bindings.protocol.draft2.processor.callback.Draft2PortProcessorHelper;
